- merge from testing-0.8.12-r0:  - tcp updates (lots)  - makefile mips support  - tm timer workarround (present also in stable), t_relay_tls changed   to t_relay_to_tls  - udp_flood sleep & throttle support

+++ fifo_server.c
@@ -349,8 +349,8 @@ static int fifo_check(int fd, char* fname)
 	 */
 	if ((lst.st_dev!=fst.st_dev)||(lst.st_ino!=fst.st_ino)){
 		LOG(L_ERR, "ERROR: security: fifo_check: inode/dev number differ"
-				": %ld %ld (%s)\n",
-				 fst.st_ino, lst.st_ino, fname);
+				": %d %d (%s)\n",
+				 (int)fst.st_ino, (int)lst.st_ino, fname);
 		return -1;
 	}
 	/* success */
